[#29] fix two bugs for pagination
Changed the listener type from select field from 'click' to 'change'. So
the listener does not fires when the select field is clicked, but when
the selected value changes.
Added a new bool keepFocus and set it to false for pagination. The input
searchfield is always set to focus after the template is rendered, now
the pagination elements do not trigger this.